Packer templates cleanup
  - fixed output_directory paths
  - corrected wait time tag in boot_command (5 sec not 5 min)

NOTE:
Still running into issues (perhaps?) between Packer products and
terraform, after successfully applying a plan when refreshing or
destroying. The VM templates' and cloned VMs' disks thin provision
and/or eager scrubbing appear to change of flip-flop boolean values
so that exceptions/errors are raised when attempting to perform
a terraform refresh and especially destroy. Not understanding yet
where this is coming from or why.